fix accept/reject job flow. fix withdraw/reapply application flow
Applicant can accept or reject as they like in offers tab. Changes will
reflect on Company page as well.

Applicant can Withdraw from a job, making the offerStatus for that job
None and disappear on the list of offers and review list for company.
Applicant can reapply, but the company has to offer the job again in the
review list since the offerStatus is None (offering it again would
change it to Pending)